+++ src/ops/mmap.cpp
@@ -6,7 +6,8 @@ extern "C" {
 void* mmap(void* addr, size_t length, int prot, int flags, int fd,
            off_t offset) {
   if (auto file = get_file(fd)) {
-    void* ret = file->mmap(addr, length, prot, flags, offset);
+    void* ret =
+        file->mmap(addr, length, prot, flags, static_cast<size_t>(offset));
     LOG_DEBUG("ulayfs::mmap(%p, %zu, %x, %x, %d, %ld) = %p", addr, length, prot,
               flags, fd, offset, ret);
     return ret;
